Show the calculated acronym.
