+++ dashboard/src/components/guilds/GuildList/GuildInfo.ts
@@ -1,0 +1,6 @@
+export interface GuildInfo {
+  id: string;
+  name: string;
+  icon?: string;
+  isPlaying: boolean;
+}
